==> 00_data_wrangling/01_first_clean.ipynb <==
# First Clean

# import library
import pandas as pd

#
#
# custom functions
#
# email reader
def read_email(email_path):
    """returns email body as a body of text"""
    
    # file manager opens email file, assigns it to variable
    with open(email_path) as email_file:
        email_body = email_file.read()
    
    # returns email body as text
    return email_body

# import working df
df = pd.read_csv('./data/00_original_wrangle.csv', header=0, index_col=False)

# check info
df.info()

# fill na with '' possibly here

# convert mime_vers to dtype object to run string methods
df.mime_vers = df.mime_vers.astype('str')


df.info()

## Start Cleaning

# PRESENTATION
# clean leading/trailing whitespace
df = df.apply(lambda x: x.str.lstrip().str.rstrip())

# PRESENTATION
# remove redundant string from f_dir
df.f_dir = df.f_dir.str.replace('./data/maildir/','')

# PRESENTATION
# returns rows where the email body is duplicated
x = len(df[df.m_body.duplicated()])

# Drop duplicates based on body text
df = df.drop_duplicates(subset = 'm_body')

# print count
print('Email Body Duplicates: {}'.format(x))

# PRESENTATION
# Select only @enron email address
only_enron = df.m_from.str.endswith('@enron.com')

# Filter df
df = df[only_enron]

# print count
print('Emails Returned  in \'@enron.com\': {}'.format(len(df)))

# PRESENTATION
# remove fowarded emails
fw_query = df.m_subj.str.startswith('FW:', na=False)

# Filter out forwarded emails
df = df[~fw_query]

# print count
print('After FW query: {}'.format(len(df)))

# USER TOOL
# email text by line as list
e_string_rev = './data/maildir/' + 'dasovich-j/all_documents/1780'
print(read_email(e_string_rev))

# Count number of emails sent per email address 

c = pd.DataFrame(df.groupby(['m_from']).size().sort_values(ascending=False))
c = c.reset_index()
c.columns = ['m_from', 'sent_counts']
c.astype({'sent_counts' : 'int'}).dtypes

# merge sent counts based on m_from email back to DataFrame
df = pd.merge(df, c, how='left', on='m_from')

# Looking at counts for 'Forwarded by' occurence
print(len(df))
print(len(df[df.m_body.str.contains('-- Forwarded by')]))

# create condition for 'Forwarded by'
cond = df.m_body.str.contains('-- Forwarded by')

# set df to ignore email body's with condition
df = df[~cond]

# print record count
print(len(df))

#
# export DataFrame for later work
#

# df.to_csv('./data/01_first_clean.csv', index=False, index_label=False)
df.head()

df.info()
